projects/exts/lead/pyproject.toml
[project]
name = "lead"
version = "0.1.0"
description = "A program that automatically gets responses from the google form and saves them to the sqlite database."
readme = "README.md"
requires-python = "==3.13.5"
dependencies = [
  "google-api-python-client==2.179.0",
  "google-auth-httplib2==0.2.0",
  "google-auth-oauthlib==1.2.2",
  "sqlalchemy==2.0.43",
]
